Refactor Notification Processors and Introduce Mock Notifs
- Refactored Email, Slack, Telegram, and Webhook notification processors to utilize dedicated services for sending notifications, improving code organization and maintainability.
- Created EmailNotificationService, SlackNotificationService, TelegramNotificationService, and WebhookNotificationService to encapsulate notification logic and error handling.
- Updated NotificationsService to support sending mock notifications for testing purposes.
- Added NotificationsController for testing webhook functionality with a public endpoint.
