+++ src/notifications/notifications.controller.ts
@@ -1,13 +1,14 @@
-import { Controller, Post, Body, Logger } from '@nestjs/common';
+import { Controller, Post, Body, Logger, Request } from '@nestjs/common';
 import { NotificationsService } from './notifications.service';
 import { Public } from 'src/auth/auth.guard';
+import { AuthenticatedRequest } from 'src/common/types/custom-types';
 
 /**
  * Controller for testing webhook functionality
  */
-@Controller('webhook')
-export class WebhookController {
-  private readonly logger = new Logger(WebhookController.name);
+@Controller('notifications')
+export class NotificationsController {
+  private readonly logger = new Logger(NotificationsController.name);
 
   constructor(private readonly notificationsService: NotificationsService) {}
 
@@ -15,7 +16,7 @@ export class WebhookController {
    * Endpoint for receiving test webhooks (use this URL as your webhook destination to test)
    * This endpoint is public to allow external services to call it
    */
-  @Post('test/receive')
+  @Post('test/webhook/receive')
   @Public()
   receiveTestWebhook(@Body() payload: any) {
     this.logger.log('Received test webhook payload');
@@ -29,5 +30,20 @@ export class WebhookController {
       message: 'Webhook received successfully',
       timestamp: new Date(),
     };
+  }
+
+  @Post('test/send')
+  sendMockNotification(
+    @Body()
+    payload: {
+      userAddress: string;
+      notificationChannel: 'webhook' | 'slack' | 'telegram' | 'email';
+    },
+    @Request() req: AuthenticatedRequest,
+  ) {
+    return this.notificationsService.sendMockNotification(
+      req.user,
+      payload.notificationChannel,
+    );
   }
 }
